+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## 2.11.0 - Unreleased
+### Added
+- `bind-route-service` command - Bind a service instance to an HTTP route
 
 ### Packaged Dependencies
 | Dependency | Version                                                             |